+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # CHANGELOG
 
 
+## v0.13.4
+* Adding `alt_text` field to image asset forms.
+* Add eme-form-block to Blocksmith.
+* cms crash to white screen after paste
+* Use `@triniti/schemas` v1.1.15
+
+
 ## v0.13.3
 * Youtube Block: Choosing start time throws error
 * Clear image for Audio & Article block not functioning on Safari.
