+++ demo/package.json
@@ -13,7 +13,7 @@
   "devDependencies": {
     "babel-loader": "^8.1.0",
     "css-loader": "^3.6.0",
-    "dotenv": "^8.1.0",
+    "dotenv": "^8.6.0",
     "file-loader": "^4.3.0",
     "node-sass": "^4.14.1",
     "postcss-loader": "^3.0.0",
@@ -21,7 +21,7 @@
     "raw-loader": "^3.1.0",
     "react-hot-loader": "^4.12.21",
     "sass-loader": "^8.0.0",
-    "sass-resources-loader": "^2.1.0",
+    "sass-resources-loader": "^2.2.1",
     "style-loader": "^1.2.1",
     "webpack": "^4.44.1",
     "webpack-cli": "^3.3.12",
