+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@triniti/cms",
-  "version": "0.13.3",
+  "version": "0.13.4",
   "description": "@triniti/cms is a single page app for managing triniti schemas and services.",
   "license": "Apache-2.0",
   "module": "./index.js",
@@ -26,13 +26,13 @@
     "@triniti/admin-ui-plugin": "^1.0.2",
     "@triniti/app": "~0.4 || ^1.0",
     "animate.css": "^3.6.1",
-    "auth0-lock": "^11.26.3",
-    "aws-sdk": "^2.740.0",
+    "auth0-lock": "^11.30.0",
+    "aws-sdk": "^2.910.0",
     "bootstrap": "^4.5.2",
     "bottlejs": "^1.7.1",
-    "chroma-js": "^2.1.0",
-    "classnames": "^2.2.5",
-    "core-js": "^3.6.5",
+    "chroma-js": "^2.1.2",
+    "classnames": "^2.3.1",
+    "core-js": "^3.12.1",
     "draft-js": "^0.11.7",
     "draft-js-buttons": "^2.0.1",
     "draft-js-export-html": "^1.2.0",
@@ -42,14 +42,14 @@
     "filesize": "^4.2.0",
     "he": "^1.1.1",
     "immutable": "^3.8.2",
-    "inflection": "^1.12.0",
+    "inflection": "^1.13.1",
     "jsondiffpatch": "^0.3.10",
     "jws": "^3.1.5",
     "jwt-decode": "^2.2.0",
-    "lodash": "^4.17.19",
-    "lodash-es": "^4.17.10",
+    "lodash": "^4.17.21",
+    "lodash-es": "^4.17.21",
     "md5": "^2.3.0",
-    "mime-types": "^2.1.27",
+    "mime-types": "^2.1.30",
     "moment": "~2.24.0",
     "mqtt": "^4.2.1",
     "prepend-http": "^3.0.1",
@@ -62,7 +62,7 @@
     "react-image-magnify": "^2.7.0",
     "react-json-tree": "^0.11.0",
     "react-player": "^1.15.3",
-    "react-redux": "^7.2.1",
+    "react-redux": "^7.2.4",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-sortable-hoc": "^1.10.1",
@@ -70,41 +70,41 @@
     "react-universal-component": "^4.0.0",
     "react-upload-button-v2": "^0.6.1",
     "react-youtube": "^7.12.0",
-    "redux": "^4.0.5",
+    "redux": "^4.1.0",
     "redux-form": "^8.3.6",
     "redux-logger": "^3.0.6",
     "redux-saga": "^1.1.1",
-    "scroll-into-view-if-needed": "^2.2.26",
+    "scroll-into-view-if-needed": "^2.2.28",
     "sortablejs": "^1.10.2",
     "sweetalert2": "^8.19.0",
     "utf8": "^3.0.0",
     "uuid": "^3.1.0",
-    "whatwg-fetch": "^3.4.1"
+    "whatwg-fetch": "^3.6.2"
   },
   "devDependencies": {
-    "@babel/cli": "^7.11.6",
-    "@babel/core": "^7.11.6",
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/plugin-proposal-decorators": "^7.10.5",
-    "@babel/plugin-proposal-export-namespace-from": "^7.10.4",
+    "@babel/cli": "^7.14.3",
+    "@babel/core": "^7.14.3",
+    "@babel/plugin-proposal-class-properties": "^7.13.0",
+    "@babel/plugin-proposal-decorators": "^7.14.2",
+    "@babel/plugin-proposal-export-namespace-from": "^7.14.2",
     "@babel/plugin-proposal-function-sent": "^7.10.4",
-    "@babel/plugin-proposal-json-strings": "^7.10.4",
-    "@babel/plugin-proposal-numeric-separator": "^7.10.4",
-    "@babel/plugin-proposal-object-rest-spread": "^7.11.0",
+    "@babel/plugin-proposal-json-strings": "^7.14.2",
+    "@babel/plugin-proposal-numeric-separator": "^7.14.2",
+    "@babel/plugin-proposal-object-rest-spread": "^7.14.2",
     "@babel/plugin-proposal-throw-expressions": "^7.10.4",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/plugin-syntax-import-meta": "^7.10.4",
-    "@babel/plugin-transform-modules-commonjs": "^7.10.4",
-    "@babel/plugin-transform-regenerator": "^7.10.4",
-    "@babel/plugin-transform-runtime": "^7.11.5",
-    "@babel/preset-env": "^7.11.5",
-    "@babel/preset-react": "^7.10.4",
-    "@babel/register": "^7.11.5",
+    "@babel/plugin-transform-modules-commonjs": "^7.14.0",
+    "@babel/plugin-transform-regenerator": "^7.13.15",
+    "@babel/plugin-transform-runtime": "^7.14.3",
+    "@babel/preset-env": "^7.14.2",
+    "@babel/preset-react": "^7.13.13",
+    "@babel/register": "^7.13.16",
     "@gdbots/schemas": "^1.6.6",
     "@hot-loader/react-dom": "^16.13.0",
     "@redux-saga/testing-utils": "^1.1.1",
-    "@triniti/acme-schemas": "^1.1.10",
-    "@triniti/schemas": "^1.1.12",
+    "@triniti/acme-schemas": "^1.1.13",
+    "@triniti/schemas": "^1.1.15",
     "ajv": "^6.12.4",
     "autoprefixer": "^9.8.6",
     "babel-eslint": "^10.1.0",
@@ -119,10 +119,10 @@
     "enzyme-adapter-react-16": "^1.15.4",
     "eslint": "^6.7.2",
     "eslint-config-airbnb": "^18.2.0",
-    "eslint-import-resolver-babel-module": "^5.1.2",
-    "eslint-plugin-import": "^2.22.0",
+    "eslint-import-resolver-babel-module": "^5.3.1",
+    "eslint-plugin-import": "^2.23.2",
     "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-react": "^7.20.6",
+    "eslint-plugin-react": "^7.23.2",
     "eslint-plugin-react-hooks": "^2.5.1",
     "faucet": "^0.0.1",
     "ignore-styles": "^5.0.1",
