+++ src/plugins/blocksmith/components/blocksmith/outer-error-boundary/styles.scss
@@ -1,0 +1,25 @@
+.blocksmith-error-boundary {
+  .invalid-block-indicator {
+    color: #f95858;
+  }
+
+  .warning {
+    color: #f99520;
+  }
+
+  .preview-component {
+    box-shadow: 0 0 0 1px #e0e0e2;
+    border-radius: 2px;
+    padding: 1rem;
+    margin-bottom: 1rem;
+
+    p {
+      margin-bottom: 0;
+    }
+  }
+
+  .preview-component__error {
+    box-shadow: 0 0 0 1px #f95858;
+    background-color: #fef1f1;
+  }
+}
