+++ src/plugins/blocksmith/components/eme-form-block-sidebar-button/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import GenericSidebarButton from '@triniti/cms/plugins/blocksmith/components/generic-sidebar-button';
+
+const config = {
+  icon: {
+    imgSrc: 'list-cards',
+  },
+};
+
+export default (props) => <GenericSidebarButton config={config} {...props} />;
