+++ src/plugins/dam/components/uploader-asset-fields/index.jsx
@@ -83,7 +83,10 @@ export default class UploaderAssetFields extends React.Component {
             <FormGroup>
               <Field name="title" component={TextField} label="Title" placeholder="Title" size="sm" />
               {asset.schema().getCurie().getMessage() === 'image-asset' && (
-                <Field name="displayTitle" component={TextField} label="Display Title" placeholder="Display Title" size="sm" />
+                <>
+                  <Field name="displayTitle" component={TextField} label="Display Title" placeholder="Display title" size="sm" />
+                  <Field name="altText" component={TextField} label="Alt Text" placeholder="Alternative information" size="sm" />
+                </>
               )}
             </FormGroup>
           </Col>
